Updated packages + gulp file
