+++ package.json
@@ -1,34 +1,33 @@
 {
-	"name": "jamisonyates.com",
-	"version": "1.0.0",
-	"description": "personal/professional website",
-	"main": "index.js",
-	"dependencies": {},
-	"devDependencies": {
-		"autoprefixer": "^9.8.5",
-		"del": "^5.1.0",
-		"gulp": "^4.0.2",
-		"gulp-csso": "^4.0.1",
-		"gulp-postcss": "^8.0.0"
-	},
-	"scripts": {
-		"build": "gulp",
-		"clean": "rm -rf dist"
-	},
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/yatrix7/jamisonyates.com.git"
-	},
-	"author": "Jamison Yates",
-	"license": "ISC",
-	"bugs": {
-		"url": "https://github.com/yatrix7/jamisonyates.com/issues"
-	},
-	"homepage": "https://github.com/yatrix7/jamisonyates.com#readme",
-	"browserslist": [
-		"> 0.5%",
-		"last 2 versions",
-		"Firefox ESR",
-		"not dead"
-	]
+  "name": "jamisonyates.com",
+  "version": "1.0.0",
+  "description": "personal/professional website",
+  "main": "index.js",
+  "devDependencies": {
+    "del": "^7.1.0",
+    "gulp": "^5.0.0",
+    "gulp-autoprefixer": "^9.0.0",
+    "gulp-csso": "^4.0.1",
+    "gulp-postcss": "^10.0.0"
+  },
+  "scripts": {
+    "build": "gulp",
+    "clean": "rm -rf dist"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/yatrix7/jamisonyates.com.git"
+  },
+  "author": "Jamison Yates",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/yatrix7/jamisonyates.com/issues"
+  },
+  "homepage": "https://github.com/yatrix7/jamisonyates.com#readme",
+  "browserslist": [
+    "> 0.5%",
+    "last 2 versions",
+    "Firefox ESR",
+    "not dead"
+  ]
 }
